+++ Engine/GameTime.h
@@ -30,12 +30,12 @@ namespace ave
     private:
 
         // Private Members
-        __int64     m_iStartTime;
-        __int64     m_iPreviousTime;
-        __int64     m_iCurrentTime;
-        __int64     m_iSecondsPerCount;
-        __int64     m_iDeltaTime;
-        __int64     m_iTotalTime;
+        float     m_iStartTime;
+        float     m_iPreviousTime;
+        float     m_iCurrentTime;
+        float     m_iSecondsPerCount;
+        float     m_iDeltaTime;
+        float     m_iTotalTime;
 
     };
 }
